Failed item pickup prompts + weight balancing
Cello and Mask weigh 8 pounds, every other item weighs 1.75. That way the player can have one of the melody items and up to 4 of the other items. Also, we limit the count of items to 5 in case the player doesn't pick a melody.

+++ MJ110Project/Assets/Scripts/Items/ItemData.cs
@@ -13,7 +13,7 @@ public class ItemData : ScriptableObject
 
     public GameObject itemPrefab;
 
-    public int Weight;
+    public float Weight;
 
     public Sprite IconSprite;
